refactor: extract new network
